added external.ts file spec (#9)
this PR adds the external.ts file spec, allowing for routing to external
applications.
The spec does not support dynamic segments in the internal path.

The tested setup runs the main app and a secondary app with a /js
basePath. The main application defines an external.ts file in
`/app/js/external.ts` exporting `http://localhost:3004` as the origin.
This results in any request starting with /js being routed to the
application on localhost:3004. All relevant middleware.ts files will be
ran before forwarding.

Internall, when an external path is matched the middleware returns
`NextResponse.next()`. Additionally the next.config is being enhanced to
add the required rewrites.